requirements.txt
altair==4.0.0
argh==0.26.2
astor==0.8.1
backcall==0.1.0
base58==1.0.3
beautifulsoup4==4.8.2
bert-extractive-summarizer==0.3.0
bleach==3.1.0
blinker==1.4
blis==0.2.4
boto3==1.10.47
botocore==1.13.47
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
idna==2.8
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.3
neuralcoref==4.0
notebook==6.0.2
numpy==1.18.1
olefile==0.46
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.0.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
pycparser==2.19
Pygments==2.5.2
python-dateutil==2.8.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
regex==2020.1.7
requests==2.22.0
s3transfer==0.2.1
sacremoses==0.0.38
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.85
six==1.13.0
soupsieve==1.9.5
spacy==2.1.3
srsly==1.0.1
streamlit==0.52.2
terminado==0.8.3
testpath==0.4.4
thinc==7.0.8
toml==0.10.0
toolz==0.10.0
torch==1.2.0
torchvision==0.4.0
tornado==5.1.1
tqdm==4.41.1
traitlets==4.3.3
transformers==2.3.0
tzlocal==2.0.0
urllib3==1.25.7
validators==0.14.1
wasabi==0.6.0
watchdog==0.9.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
